Partially Working Version (v0.1)
This is the first presentable version of the HUD. However some work still needs to be done, such as scaling/repositioning for class and weapon specific hud elements and fine tuning the character portrait for each class. However it should be completely presentable for Scout and usable and mostly fine for classes other than Medic, especially if only using stock weapons.